+++ examples/preemptible_and_regular_instance_templates/simple/versions.tf
@@ -14,12 +14,6 @@
  * limitations under the License.
  */
 
-output "self_link" {
-  description = "Self-link of instance template"
-  value       = google_compute_instance_template.tpl.self_link
-}
-
-output "name" {
-  description = "Name of instance template"
-  value       = google_compute_instance_template.tpl.name
+terraform {
+  required_version = ">= 0.12"
 }