+++ modules/preemptible_and_regular_instance_templates/outputs.tf
@@ -14,12 +14,22 @@
  * limitations under the License.
  */
 
-output "self_link" {
-  description = "Self-link of instance template"
-  value       = google_compute_instance_template.tpl.self_link
+output "preemptible_self_link" {
+  description = "Self-link of preemptible instance template"
+  value       = module.preemptible.self_link
 }
 
-output "name" {
-  description = "Name of instance template"
-  value       = google_compute_instance_template.tpl.name
+output "preemptible_name" {
+  description = "Name of preemptible instance template"
+  value       = module.preemptible.name
+}
+
+output "regular_self_link" {
+  description = "Self-link of regular instance template"
+  value       = module.regular.self_link
+}
+
+output "regular_name" {
+  description = "Name of regular instance template"
+  value       = module.regular.name
 }