+++ CHANGELOG.md
@@ -7,12 +7,14 @@ project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
 
+## [1.1.0] - 2019-10-23
+
 ### Added
 
 - Added variable `project_id`. [#26]
-
 - `preemptible_and_regular_instance_templates` submodule. [#18]
-- `mig_with_percent` submodule. [#18]
+- `mig_with_percent` submodule. [#19]
+- Support for shielded VMs. [#38]
 
 ## [1.0.0] - 2019-07-31
 
@@ -36,7 +38,8 @@ project adheres to [Semantic Versioning](http://semver.org/).
 - `instance_template`, `mig`, and `umig` modules
 - examples, basic tests and fixtures for all modules
 
-[Unreleased]: https://github.com/terraform-google-modules/terraform-google-vm/compare/v1.0.0...HEAD
+[Unreleased]: https://github.com/terraform-google-modules/terraform-google-vm/compare/v1.1.0...HEAD
+[1.1.0]: https://github.com/terraform-google-modules/terraform-google-vm/compare/v1.0.0...v1.1.0
 [1.0.0]: https://github.com/terraform-google-modules/terraform-google-vm/compare/v0.2.0...v1.0.0
 [0.2.0]: https://github.com/terraform-google-modules/terraform-google-vm/compare/v0.1.0...v0.2.0
 [0.1.0]: https://github.com/terraform-google-modules/terraform-google-vm/releases/tag/v0.1.0
@@ -44,4 +47,6 @@ project adheres to [Semantic Versioning](http://semver.org/).
 [#14]: https://github.com/terraform-google-modules/terraform-google-vm/pull/14
 [#16]: https://github.com/terraform-google-modules/terraform-google-vm/pull/16
 [#18]: https://github.com/terraform-google-modules/terraform-google-vm/pull/18
+[#19]: https://github.com/terraform-google-modules/terraform-google-vm/pull/19
 [#26]: https://github.com/terraform-google-modules/terraform-google-vm/pull/26
+[#38]: https://github.com/terraform-google-modules/terraform-google-vm/pull/38